requirements.txt
pillow==11.0.0
python-dotenv==1.1.1
requests~=2.32.5
colorlog~=6.9.0